cmd/requirements.txt
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.1.0
